requirements.txt
# Django
Django>=5.1,<5.2
djangorestframework>=3.15.0
django-cors-headers>=4.3.0

# Database
dj-database-url>=2.1.0

# Docker SDK
docker>=7.0.0

# Stripe
stripe>=8.0.0

# HTTP requests (for health checks)
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Production server
gunicorn>=21.0.0

# For generating secure keys
secrets
